requirements.txt
Flask
mysql-connector-python
scikit-learn
pandas
numpy
joblib
nltk
